:zap: Move user-item pairs to index of dataframe
Improve performance of TradesyDataset by using both user and item IDs
to create a `MultiIndex`. The performance improvement comes from the
lookup of the candidate negative item (lookup on a MultiIndex is
faster than using `.loc` in a dataframe). This required changing some
parts of the custom dataset, switching from extensive usage of `.loc`
and `.iloc` to using `get_level_values` on the `MultiIndex` when
necessary

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>