:sparkles: Refactor recommendation API
Instead of using multiple methods, unify predictions in a single
method. To ensure that the interface is understandable, the method
will raise useful error messages when necessary, and use a similar
interface to the asterisk (*) operator. If two tensors can be operated
as users * items, then both are acceptable inputs for this function

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>